+++ tools/xmlGenerator/CMakeLists.txt
@@ -28,9 +28,12 @@
 
 install(PROGRAMS
     domainGenerator.sh
+    domainGenerator.py
     hostConfig.py
     hostDomainGenerator.sh
     lightRoutingUpdate.sh
+    PfwBaseTranslator.py
+    EddParser.py
     PFWScriptGenerator.py
     portAllocator.py
     updateRoutageDomains.sh
